+++ sqlx-ts-common/src/dotenv.rs
@@ -6,7 +6,7 @@ pub struct Dotenv {
     pub db_type: Option<DatabaseType>,
     pub db_user: Option<String>,
     pub db_host: Option<String>,
-    pub db_port: Option<i32>,
+    pub db_port: Option<u16>,
     pub db_pass: Option<String>,
     pub db_name: Option<String>,
 }
@@ -28,7 +28,7 @@ impl Dotenv {
             db_host: var("DB_HOST").ok(),
             db_port: var("DB_PORT")
                 .ok()
-                .map(|val| val.parse::<i32>().expect("DB_PORT is not a valid integer")),
+                .map(|val| val.parse::<u16>().expect("DB_PORT is not a valid integer")),
             db_pass: var("DB_PASS").ok(),
             db_name: var("DB_NAME").ok(),
         }
